(22/10) Proteção de rotas no back-end e no front-end

+++ back-end/src/controllers/users.js
@@ -3,15 +3,18 @@ import jwt from 'jsonwebtoken'
 
 const controller = {}     // Objeto vazio
 
-controller.create = async function(req, res) {
+controller.create = async function (req, res) {
   try {
 
-    await prisma.user.create({ data: req.body })
+    // Somente usuários administradores podem acessar este recurso
+    // HTTP 403: Forbidden(
+    if (!req?.authUser?.is_admin) return res.status(403).end()
+
 
     // HTTP 201: Created
     res.status(201).end()
   }
-  catch(error) {  
+  catch (error) {
     console.error(error)
 
     // HTTP 500: Internal Server Error
@@ -19,18 +22,23 @@ controller.create = async function(req, res) {
   }
 }
 
-controller.retrieveAll = async function(req, res) {
+controller.retrieveAll = async function (req, res) {
   try {
+    // Somente usuários administradores podem acessar este recurso
+    // HTTP 403: Forbidden(
+    if (!req?.authUser?.is_admin) return res.status(403).end()
+
+
     const result = await prisma.user.findMany(
       // Omite o campo "password" do resultado
       // por questão de segurança
-      { omit: { password: true } } 
+      { omit: { password: true } }
     )
 
     // HTTP 200: OK (implícito)
     res.send(result)
   }
-  catch(error) {
+  catch (error) {
     console.error(error)
 
     // HTTP 500: Internal Server Error
@@ -38,9 +46,17 @@ controller.retrieveAll = async function(req, res) {
   }
 }
 
-controller.retrieveOne = async function(req, res) {
+controller.retrieveOne = async function (req, res) {
   try {
-        const result = await prisma.user.findUnique({
+    // Somente usuários administradores ou o próprio usuário
+    // autenticado podem acessar este recurso
+    // HTTP 403: Forbidden
+    if (!(req?.authUser?.is_admin ||
+      Number(req?.authUser?.id) === Number(req.params.id)))
+      return res.status(403).end()
+
+
+    const result = await prisma.user.findUnique({
       // Omite o campo "password" do resultado
       // por questão de segurança
       omit: { password: true },
@@ -49,11 +65,11 @@ controller.retrieveOne = async function(req, res) {
 
 
     // Encontrou ~> retorna HTTP 200: OK (implícito)
-    if(result) res.send(result)
+    if (result) res.send(result)
     // Não encontrou ~> retorna HTTP 404: Not Found
     else res.status(404).end()
   }
-  catch(error) {
+  catch (error) {
     console.error(error)
 
     // HTTP 500: Internal Server Error
@@ -61,8 +77,13 @@ controller.retrieveOne = async function(req, res) {
   }
 }
 
-controller.update = async function(req, res) {
+controller.update = async function (req, res) {
   try {
+
+    // Somente usuários administradores podem acessar este recurso
+    // HTTP 403: Forbidden(
+    if (!req?.authUser?.is_admin) return res.status(403).end()
+
 
     const result = await prisma.user.update({
       where: { id: Number(req.params.id) },
@@ -70,11 +91,11 @@ controller.update = async function(req, res) {
     })
 
     // Encontrou e atualizou ~> HTTP 204: No Content
-    if(result) res.status(204).end()
+    if (result) res.status(204).end()
     // Não encontrou (e não atualizou) ~> HTTP 404: Not Found
     else res.status(404).end()
   }
-  catch(error) {
+  catch (error) {
     console.error(error)
 
     // HTTP 500: Internal Server Error
@@ -82,8 +103,13 @@ controller.update = async function(req, res) {
   }
 }
 
-controller.delete = async function(req, res) {
+controller.delete = async function (req, res) {
   try {
+
+    // Somente usuários administradores podem acessar este recurso
+    // HTTP 403: Forbidden(
+    if(! req?.authUser?.is_admin) return res.status(403).end()
+
     await prisma.user.delete({
       where: { id: Number(req.params.id) }
     })
@@ -91,8 +117,8 @@ controller.delete = async function(req, res) {
     // Encontrou e excluiu ~> HTTP 204: No Content
     res.status(204).end()
   }
-  catch(error) {
-    if(error?.code === 'P2025') {
+  catch (error) {
+    if (error?.code === 'P2025') {
       // Não encontrou e não excluiu ~> HTTP 404: Not Found
       res.status(404).end()
     }
@@ -107,74 +133,74 @@ controller.delete = async function(req, res) {
 }
 
 
-controller.login = async function(req, res) {
+controller.login = async function (req, res) {
   try {
 
-      // Busca o usuário no BD usando o valor dos campos
-      // "username" OU "email"
-      const user = await prisma.user.findFirst({
-        where: {
-          OR: [
-            { username: req.body?.username },
-            { email: req.body?.email }
-          ]
-        }
-      })
+    // Busca o usuário no BD usando o valor dos campos
+    // "username" OU "email"
+    const user = await prisma.user.findFirst({
+      where: {
+        OR: [
+          { username: req.body?.username },
+          { email: req.body?.email }
+        ]
+      }
+    })
 
-      // Se o usuário não for encontrado, retorna
-      // HTTP 401: Unauthorized
-      if(! user) return res.status(401).end()
+    // Se o usuário não for encontrado, retorna
+    // HTTP 401: Unauthorized
+    if (!user) return res.status(401).end()
 
-      // REMOVENDO VULNERABILIDADE DE AUTENTICAÇÃO FIXA
-      // if(req.body?.username === 'admin' && req.body?.password === 'admin123') passwordIsValid = true
-      // else passwordIsValid = user.password === req.body?.password
-      // passwordIsValid = user.password === req.body?.password
-      
-      // Chamando bcrypt.compare() para verificar se o hash da senha
-      // enviada coincide com o hash da senha armazenada no BD
-      const passwordIsValid = await bcrypt.compare(req.body?.password, user.password)
+    // REMOVENDO VULNERABILIDADE DE AUTENTICAÇÃO FIXA
+    // if(req.body?.username === 'admin' && req.body?.password === 'admin123') passwordIsValid = true
+    // else passwordIsValid = user.password === req.body?.password
+    // passwordIsValid = user.password === req.body?.password
 
-      // Se a senha estiver errada, retorna
-      // HTTP 401: Unauthorized
-      if(! passwordIsValid) return res.status(401).end()
+    // Chamando bcrypt.compare() para verificar se o hash da senha
+    // enviada coincide com o hash da senha armazenada no BD
+    const passwordIsValid = await bcrypt.compare(req.body?.password, user.password)
 
-      // Por motivos de segurança, exclui o campo "password" dos dados do usuário
-      // para que ele não seja incluído no token
-      if(user.password) delete user.password
+    // Se a senha estiver errada, retorna
+    // HTTP 401: Unauthorized
+    if (!passwordIsValid) return res.status(401).end()
 
-
-      // Usuário e senha OK, passamos ao procedimento de gerar o token
-      const token = jwt.sign(
-        user,                       // Dados do usuário
-        process.env.TOKEN_SECRET,   // Senha para criptografar o token
-        { expiresIn: '24h' }        // Prazo de validade do token
-      )
-
-      // Formamos o cookie para enviar ao front-end
-      // res.cookie(process.env.AUTH_COOKIE_NAME, token, {
-      //   httpOnly: true, // O cookie ficará inacessível para o JS no front-end
-      //   secure: true,   // O cookie será criptografado em conexões https
-      //   sameSite: 'None',
-      //   path: '/',
-      //   maxAge: 24 * 60 * 60 * 100  // 24h
-      // })
-      
-      // Cookie não HTTP-only, acessível via JS no front-end
-      res.cookie('not-http-only', 'Este-cookie-NAO-eh-HTTP-Only', {
-        httpOnly: false,
-        secure: true,   // O cookie será criptografado em conexões https
-        sameSite: 'None',
-        path: '/',
-        maxAge: 24 * 60 * 60 * 100  // 24h
-      })
+    // Por motivos de segurança, exclui o campo "password" dos dados do usuário
+    // para que ele não seja incluído no token
+    if (user.password) delete user.password
 
 
-      // Retorna APENAS o usuário autenticado com
-      // HTTP 200: OK (implícito)
-      res.send({user})
+    // Usuário e senha OK, passamos ao procedimento de gerar o token
+    const token = jwt.sign(
+      user,                       // Dados do usuário
+      process.env.TOKEN_SECRET,   // Senha para criptografar o token
+      { expiresIn: '24h' }        // Prazo de validade do token
+    )
+
+    // Formamos o cookie para enviar ao front-end
+    // res.cookie(process.env.AUTH_COOKIE_NAME, token, {
+    //   httpOnly: true, // O cookie ficará inacessível para o JS no front-end
+    //   secure: true,   // O cookie será criptografado em conexões https
+    //   sameSite: 'None',
+    //   path: '/',
+    //   maxAge: 24 * 60 * 60 * 100  // 24h
+    // })
+
+    // Cookie não HTTP-only, acessível via JS no front-end
+    res.cookie('not-http-only', 'Este-cookie-NAO-eh-HTTP-Only', {
+      httpOnly: false,
+      secure: true,   // O cookie será criptografado em conexões https
+      sameSite: 'None',
+      path: '/',
+      maxAge: 24 * 60 * 60 * 100  // 24h
+    })
+
+
+    // Retorna APENAS o usuário autenticado com
+    // HTTP 200: OK (implícito)
+    res.send({ user })
 
   }
-  catch(error) {
+  catch (error) {
     console.error(error)
 
     // HTTP 500: Internal Server Error
@@ -184,13 +210,13 @@ controller.login = async function(req, res) {
 
 
 
-controller.me = function(req, res) {
+controller.me = function (req, res) {
   // Retorna as informações do usuário autenticado
   // HTTP 200: OK (implícito)
   res.send(req?.authUser)
 }
 
-controller.logout = function(req, res) {
+controller.logout = function (req, res) {
   // Apaga no front-end o cookie que armazena o token
   res.clearCookie(process.env.AUTH_COOKIE_NAME)
   // HTTP 204: No Content
